README.md
Creating a GitHub Repository* 

Signed in to GitHub account 
Create a New repository
Name repository Hello World 
Make the repository public 

Create and Save Python Script
Add File to your git hub 

*Challenges *
It was a bit confusing at first 

*Summary* 

Using GitHub for this assingment was a bit complex. The process was a bit confusing first but it worked as expected.